backend/requirements.txt
<<<<<<< HEAD
=======
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask==3.0.3
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
Pygments==2.18.0
SQLAlchemy==2.0.31
Werkzeug==3.0.3
>>>>>>> main
alembic==1.13.1
appnope==0.1.4
asttokens==2.4.1
blinker==1.8.2
branca==0.7.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
<<<<<<< HEAD
contourpy==1.2.1
cramjam==2.8.3
cycler==0.12.1
=======
cramjam==2.8.3
>>>>>>> main
debugpy==1.8.1
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
<<<<<<< HEAD
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
folium==0.17.0
fonttools==4.53.0
=======
folium==0.17.0
>>>>>>> main
fsspec==2024.6.1
googlemaps==4.10.0
greenlet==3.0.3
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
<<<<<<< HEAD
kiwisolver==1.4.5
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.0
=======
>>>>>>> main
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.3
numpy==2.0.0
openpyxl==3.1.4
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
<<<<<<< HEAD
Pygments==2.18.0
pyparsing==3.1.2
=======
>>>>>>> main
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.0.3
requests==2.32.3
scipy==1.14.0
seaborn==0.13.2
seirsplus==1.0.9
six==1.16.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
xyzservices==2024.6.0
